+++ NEWS.md
@@ -1,6 +1,9 @@
 # Version 4.11.1
 2018-mm-dd
 
+Announcements:
+ - pg-mvt: Fix bug while building query and there is no columns defined for the layer.
+
 
 # Version 4.11.0
 2018-10-15